Add files via upload
Added updated main method with an option to create a new account/profile, load an existing profile, and exit the program. All of which will be located in a new mysocialprofile.txt file. 

I also added two new methods for login: 
1. loadProfile()
where the user can load an existing profile and sign in from there 
2. createAccount()
where the user can create a new account with a username, password, email, and class year. This is the option I showed you both yesterday, but I still want to make the email need the "@" symbol and the class year to be a 4 digit number that correlates to the current year.